backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-auth
pyjwt
langchain
langchain-community
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
pypdf
requests
authlib
starlette
trafilatura
requests-cache
python-multipart
aiohttp
tqdm
pdfplumber
pymupdf
pytesseract
Pillow
duckduckgo-search
openai
beautifulsoup4
